requirements.txt
pymupdf==1.25.1
anthropic==0.42.0
spacy==3.8.3
tenacity==9.0.0
python-dotenv==1.0.1
msgspec==0.19.0
flask>=3.0.0
